requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
